Moved app.js and server.js to src
