+++ backend/src/server.js
@@ -1,7 +1,7 @@
 const { PORT = 5001 } = process.env;
 
 const app = require("./app");
-const knex = require("./src/db/connection");
+const knex = require("./db/connection");
 
 knex.migrate
   .latest()
